requirements.txt
python-dotenv~=0.20.0
RPi.GPIO2
requests==2.28.0